add simple HIKe Testbed
add very simple testbed in "testbed" folder with the purpose of showing
the entire HIKe Toolchain developed so far. Please *READ* the
testbed/READ.me file to understand how to interact with the testbed.

It is *highly* recommended to study testbed/testbed.sh for a deep
understanding of how the various 'components' of HIKe fit together.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>