add external Makefile for compiling eBPF HIKe Progs and HIKe Chains
external Makefile allows us to compile eBPF HIKe Progs (.bpf.c) and HIKE
Chains (.hike.c) from an 'out-of-tree' directory.

Artifacts are automatically placed in the PWD under the 'build'
directory.
---

Hereafter an example on how to call the Makefile contained in external/
directory.

For compiling an eBPF/HIKe Program:

 $ make -f path-to/hike_vm/external/Makefile -j24 prog PROG=prog.bpf.c HIKE_DIR=path-to/hike_vm/src/

where you need to specify the 'prog' target and program name through the
'PROG' variable. HIKE_DIR must point to the 'src' folder of the HIKe VM
project.

For compiling an HIKe Chain:

 $ make -f path-to/hike_vm/external/Makefile -j24 chain CHAIN=chain.hike.c HIKE_DIR=path-to/hike_vm/src/

This time the target is 'chain' while the variable which refers to the
chain to be compiled is 'CHAIN'.
HIKE_DIR must point to the 'src' folder of the HIKe VM project.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>