update testbed on tail calls
The testbed for tail calls has been revised. In particular, in the SUT we
find hooked on the enp6s0f0 interface the loader for testing tail calls in
raw eBPF, while on the enp6s0f1 interface we find the loader to execute the
tail calls in HIKe.

The purpose of the test is to calculate the performance of the HIKe VM
while executing tail calls compared to the raw eBPF approach.
To do this, we designed a (logical) program that performs 4 tail calls: a)
3 tail calls simulate the same operation on the packet (in reality, we do
nothing on the packet); b) 1 tail call performs a cross-connection between
the packet's ingress interface and the programmed outgoing interface.
The relationship between input/output interfaces is dynamically programmed
in a map.

This program was implemented in eBPF raw (i.e., without the help of the
HIKe VM), and in HIKe.
As for HIKe, basic HIke eBPF programs such as the one that simulates a
packet operation and the HIKe eBPF program to perform cross-connection
were implemented.

Finally, the logic by which these programs interact is captured in a
specific HIKe Chain that is loaded by an HIKe loader and executed upon the
arrival of a packet on the enp6s0f1 interface.

For more details please look at the code.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>