add support for exporting HIKe maps in loader/classifier program
loader/classifier is not a full HIKe eBPF program. Indeed, it is used to
boostrap the HIKe VM by selecting the right chain to kick in.
For this reason, a loader/classifer should NOT use the
EXPORT_HIKE_PROG_MAP(hikeprogname, mapname) macro, but instead we should
use the new macro EXPORT_HIKE_MAP(progname, mapname).

*DO NOT* USE EXPORT_HIKE_MAP in an HIKe eBPF program and vice versa
(i.e.: *DO NOT* USE EXPORT_HIKE_PROG_MAP in a loader/classifier).

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>