add support for DEBUG_HKPRG_PRINT function
the DEBUG_HKPRG_PRINT macro allows the user to print out the name of the
HIKe eBPF Program in the bpf trace.

This behavior can be enabled by defining the HIKE_PROG_NAME macro with
the name of the HIKe eBPF Program name, i.e.:

 /* !!! NOTE: it must come before ANY #include !!! */
 #define HIKE_PROG_NAME my_hike_prog

 /* any include here */
 #include ...

 HIKE_PROG(HIKE_PROG_NAME)
 {
 	...

	DEBUG_HKPRG_PRINT("Hello world");
	...
 }

In the bpf trace pipe:
       ...
       <idle>-0  [019] d.s. 9957329.153395: bpf_trace_printk: my_hike_prog: Hello world
       ...

If the HIKE_PROG_NAME is not defined, the DEBUG_HKPRG_PRINT behaves as
the DEBUG_PRINT macro.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>