change Chain ID numeration schema
so far, chain IDs were starting from 0x40 (64 dec). This numeration
schema was only for testing and debug purposes.

In this patch, each HIKe VM Chain ID or HIKe eBPF Program ID (for short
HIKe Program ID) is represented using a 32-bit word (dword).

To discriminate a Chain ID from a Program ID, the 30-th bit of the ID
(counting from the LSB at 0 index) is used.
When the 30-th bit is SET (=1) the ID represents a valid Chain ID,
otherwise the ID has to be considerd as an HIKe Program.

 MSB              LSB
  31  30 ... 2  1  0  | Bit i-th
  --------------------+
   0   y ... c  b  a  | ID value

 if y == 0 then ID is an HIKe Program valid ID
 if y == 1 then ID is a Chain valid ID

NOTE: the 31-th MUST be set to 0 in any case.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>