add the HIKE_VM_VERCOMP_LEVEL for the HIKe VM
in order to improve compatibility for verifiers (and programs with
verifiers), the HIKE_VM_VERCOMP_LEVEL macro can be set in the HIKe VM.

When the HIKE_VM_VERCOMP_LEVEL > 0, the relax_verifier() is compiled
within the HIKe VM code.

If the HIKE_VM_VERCOMP_LEVEL is not defined by the user, the HIKe VM
assumes its value set to 0.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>