set the HIKE_VM_VERCOMP_LEVEL to 1 by default
it seems that with many insns per HIKe Chain the verifier complains about
the HIKe VM code. To fix that we turn on the compatibility level
(HIKE_VM_VERCOMP_LEVEL to 1) so that the relax_verifier() is turned on
automatically.

TODO: we can set the HIKE_VM_VERCOMP_LEVEL on the basis of the number of
insns supported by the HIKe VM.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>