find IPv6 extension header protocol or final protocol
find the offset to specified header or the protocol number of last
header if target < 0. "last header" is transport protocol header, ESP,
or "No next header".

Note that *offset is used as input/output parameter, and if it is not
zero, then it must be a valid offset to an inner IPv6 header. This can
be used to explore inner IPv6 header, eg. ICMPv6 error messages.
If *offset is zero, then the header cursor Network Offset
(hdr_cursor->nhoff) must be a valid offset pointing to the IPv6
Header.

If target header is found, its offset is set in *offset and return
protocol number. Otherwise, return -ENOENT.

If the first fragment doesn't contain the final protocol header or
NEXTHDR_NONE it is considered invalid.

IP6_FH_F_AUTH flag is set and target < 0, then this function will
stop at the AH header.
If IP6_FH_F_SKIP_RH flag was passed, then this function will skip all
those
routing headers, where segments_left was 0.

---

NOTE: fragment header is not supported yet.

NOTE: while searching for target header, the max depth (number of
examined protocols) is set to IPV6_EXTHDR_DEPTH_MAX. If the max depth
limit is reached before the target header is found, then return
-ELOOP.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>