add support for accessing HIKe VM per-CPU Shared Memory at given offset
the hike_pcpu_shmem_obj helper function allows the user to access the
shared memory for loading/storing any kind of object.

Hereafter an example:

HIKE_PROG(MYPROG)
{
 ...

 struct __shm_buff {
   char p[BUF_LEN];
 } *pshm;

 ...

 /* reserve some space for storing the string to be searched */
 pshm = hike_pcpu_shmem_obj(sizeof(struct pkt_info), struct __shm_buff);
 if (unlikely(!pshm))
         goto abort;

 /* set the string in the shmem, so we do not overload the stack.
  * In this case, the keyword to be found is pretty small and then it
  * can be placed into the stack, directly.
  * Howerver, this exmaple shows a possible to way for loading very long
  * strings or huge data block withouth hogging the stack (<= 512 bytes).
  */
 pshm->p[0] = 'f';
 pshm->p[1] = 'o';
 pshm->p[2] = 'o';

 ...
}

the hike_pcpu_shem_obj(offset, obj) returns a pointer to the shared
memory starting from the offset value. The obj value is the type of the
object (struct, scalar, etc) that should be read/written from/to.

In the example above, the offset is set to sizeof(struct pkt_info) since
the first sizeof(struct pkt_info) bytes (for short X bytes) are occupied
by the pkt_info structure. Therefore, the phsm pointer is set after X
bytes since the beginning of the per-CPU Shared Memory. The type of the
phsm is struct __shm buff.

If there is no space enough in the shared memory for containing the type
object, the hike_pcpu_shmem_obj returns NULL; otherwise it returns a
valid pointer.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>