add barrier to avoid compiler/optimizer insns reordering
the hike_pcpu_shmem_obj(off, size) macro is used to reserve 'size' bytes of
memory on the per-CPU Shared Memory at offset 'off'. The macro returns
the pointer to that memory in case of success, NULL otherwise.
Therefore, we must check for the returned pointer value to avoid any
NULL dereferencing.

However under some circumstances, the compiler/optimizer can
reorganize insns order in a way that the verifier complains about NULL
derefercing/invalid operations.
In this case, the base address of the shared memory map is returned and
the 'off' is added to the base pointer as displacement. Only after
that, the base pointer is checked for any NULL value.
That causes a reject from the verifier which complains about prohibited
pointer arithmetic with map/NULL pointers.

To avoid insns reordering, a barrier() has been added so that any check
on the returned pointer is evaluated before performing any aritmethic
operation on that pointer.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>