add srh inline pop capability
this is just a showcase with the purpose of demonstrating how to pop an
header from a packet: in this case we pop the inline SRH.
Note that after the popping operation, the hdr_cursor offsets shold not
be considered valid anymore. In other words, they must be evaluated once
again by parsing the packet accordingly to the specific use case.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>