add memmove helper function for HIKe eBPF Programs
the builtin 'memmove' doesn't like to deal with lengths stored in variables
whose values are evaluated dynamically.
This implementation of memmove is useful when we need to perform pop
operations on protocol headers which are not fixed in size.

The current implementation of hike_memmove is able to handle up to 511
bytes at a time.
The function returns 0 on success, < 0 otherwise.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>