reduce stack pressure for ipv6_udp_checksum
in ipv6_udp_checksum, partial checksum on IPv6 addresses are evaluated
using the ipv6_addr_sum(...).
The compiler emits machine code for the ipv6_addr_sum so that it
stresses the use of eBPF stack. In other terms, the compiler produces
machine code which heavily uses stack space.
This is not good if we want to keep the stack overhead as little as
possible.

To overcome this issue, the ipv6_addr_sum makes use of some barrier
compiler tricks. In this way, the partial checksum on the IPv6 is done
using only a couple of registers and the result is pushed into the stack
consuming only one memory location (size of __be32 in bytes).

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>