add cur_xdp_adjust_head helper function
The cur_xdp_adjust_head(...) helper function allows the user to adjust
the xdp frame head and to keep in sync the offsets in the header cursor
(hdr_cursor).

This helper function is useful for shrinking or expanding the xdp frame
head (using bpf_xdp_adjust_head) when encap/encap operations on a packet
are needed.

The cur_xdp_adjust_head(struct xdp_md *ctx, struct hdr_cursor *cur,
int off) helper function takes 3 arguments:
  - ctx: xdp context;
  - cur: the current header cursor;
  - off: number of bytes to shrink or expand in the xdp frame head.

The sign of 'off' argument decides if the xdp frame head should be
shrunk or expanded.
If off < 0, the xdp frame head is expanded of "off" bytes; conversely,
if off >= 0, the xdp frame head is shrunk of "off" bytes.
Offsets of header cursor are adjusted according to the "-off" value.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>