improve the generic SRH inline pop operation
we make use of cur_xdp_adjust_head(...) for shrinking the xdp frame head
and keeping in synch the header cursor offsets.

After the SRH inline pop operation we also parse the moved headers once
again so that header {mhoff, nhoff} are valid and correctly point to
the mac layer and to the IPv6 header.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>