add support for SRv6 Encap in HIKe
HIKe supports SRv6 H.Encaps which allows IPv4/IPv6 packets to be
encapsulated in an outer IPv6 plush SRH.

Currently, HIKe SRv6 H.Encaps support is provided through the
hike_sr6_encap.bpf.c file. Such a HIKe eBPF program can be used in HIKe
chains and is not a "terminal" program, i.e., it returns the control
flow to the HIKe VM.
SRv6 policies can be inserted into eBPF sr6_encap_policy_X maps, where X
is between 1 and 8. X represents the number of SIDs of which the SID
List is composed, i.e., the specific policy.
Each policy consists of a "fixed" part of the SRH (8 bytes) and the SID
List. Therefore, the user is responsible for providing SRv6 Policy
properly composed as a whole.

The source address (SA) of the external IPv6 header is set using the
entry at index 0 of the sr6_src_tun eBPF map. This entry must be set by
the user space and its validity is for the entire network namespace,
i.e., the SA address is used for each SRv6 encapsulation within the same
network namespace.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>