add support for RAW IPv6 packet forwarding accelerator
This eBPF program aims to accelerate the forwarding of IPv6 packets once
egress network device and neighbor information are collected. Such data
is stored in the eBPF map named lookup ip6_fwd_map.
For each incoming
packet, the program looks up in the map, and if the IPv6 DA matches one
of the entries, the packet is forwarded directly to the neighbor
(realizing the mac rewrite procedure).

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>