add support for new the clssifier ip46_simple_classifier
the new ip46_simple_classifier is able to classify packets based on the
L3 protocol, i.e., IPv4 and IPv6. Based on the L3 protocol, the
classifier is able to execute the properly configured HIKe chain.

The ip46_simple_classifier can be programmed through the eBPF
ip46_simp_cls_map. Specifically:
 -  the map entry at index 1 can be set with the HIKe chain ID
    that is executed when the classifier handles IPv4 packets;
 -  the map entry at index 2 can be configured with the HIKe chain ID to
    be executed when the classifier processes IPv6 packets.

The entry at index 0 should not be used and is reserved for internal
use.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>