add support for IPv6/SRH/UDP(dest=862) traffic classifier
This classifier is introduced to provide an efficient way to filter
IPv6/SRH/UDP packets where the UDP destination port is set to 862.

As with ip6_simpler_classifier, the classifier must be programmed by
setting the chain ID in the eBPF ip6_udport_cls_map (map entry at index
0). Once loaded, the classifier can be attached (XDP hook) to the
network device where the traffic is intended to be received.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>