add helper functions for type-casting HIKe VM registers
Simple helper functions have been introduced for type-casting the HIKe
VM registers.

The HVM_PROG_RET(retcode, action) macro is introduced to simplify the
returning path from an HIKe Program to the caller HIKe chain:
 - i)  by notifing to the HIKe VM the action be taken (e.g., XDP_PASS,
       XDP_DROP, etc);
 - ii) by setting the register REG0 (aka HVM_RET) with the return code
       that the program wants to give back to the caller chain.

It is highly recommended to use this helper function instead of passing
the return code to the caller chain AND the action to the HIKe VM
separately.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>