improve the sr6_encap HIKe Program
The sr6_encap HIKe Program takes two args:
  i) __u8 nsids: the number of segments of which the Policy is composed;
 ii) __u32 index: the index of the Policy to be selected in the SRv6
     Policy eBPF map made of @nsids segments.

The program returns an 'int' to the calling chain indicating whether the
encap operation was successful or not. If successful, the ret code is 0;
otherwise the ret code is < 0. Specifically, the ret code is -ENOENT
when there are no policies that match the <nsids,index> values provided
to this program. This turns out to be a recoverable error, and
execution of the chaena continues.

Conversely, if the encapsulation operation fails for other reasons, this
error is considered unrecoverable and the packet is dropped.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>