force reg_val variable to be always initialized during HIKe VM ALU ops
with clang/llvm version 14, the HIKe VM code cannot be successfully
verified. This is due to the way the compiler/optimizer reorganizes the
produced bytecode.
The problem seems to be related to an invalid access (read before init)
to the 'reg_val' value used in the ALU operations of HIKe VM.
To solve this problem, the 'reg_val' value is always set to 0 before
performing any ALU operation, although that value can be overwritten
immediately afterwards by the specific operations.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>