force testbeds to use provided statically-linked bpftool
testbeds have been updated to use the statically-linked bpftool provided
within this repository. This is due to the fact that new versions of
bpftool handle the eBPF program names differently from each other.
Specifically, new versions of bpftool pin eBPF programs using the
function names instead of the section names. In this way, testbeds can
be run without depending on the version of the bpftool available in the
environment.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>